requirements.txt
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyascii==0.3.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
boto3==1.26.135
botocore==1.29.135
bpemb==0.3.4
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
conllu==4.5.2
contourpy==1.0.7
contractions==0.1.73
cycler==0.11.0
datasets==2.12.0
Deprecated==1.2.13
dill==0.3.6
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
django-redis==5.2.0
djangorestframework==3.14.0
emoji==2.2.0
fake-useragent==1.1.3
filelock==3.12.0
flake8==6.0.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.3
gdown==4.4.0
gensim==4.3.1
googlesearch-python==1.2.3
huggingface-hub==0.14.1
hyperopt==0.2.7
idna==3.4
inquirerpy==0.3.4
isort==5.12.0
Janome==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
langdetect==1.0.9
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
more-itertools==9.1.0
mpld3==0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
pfzy==0.3.4
Pillow==9.5.0
platformdirs==3.5.1
pptree==3.1
prompt-toolkit==3.0.38
protobuf==3.20.2
psutil==5.9.5
psycopg2==2.9.6
py4j==0.10.9.7
pyahocorasick==2.0.0
pyarrow==12.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pyre-extensions==0.0.23
PySocks==1.7.1
python-dateutil==2.8.2
pytorch_revgrad==0.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.30.0
responses==0.18.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
sqlitedict==2.1.0
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
textsearch==0.0.24
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tqdm==4.65.0
transformer-smaller-training-vocab==0.2.3
transformers==4.29.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Wikipedia-API==0.5.8
wrapt==1.15.0
xformers==0.0.16
xxhash==3.2.0
yarl==1.9.2
